localEr

+++ novelSnvFilter_ACGT.cpp
@@ -657,7 +657,7 @@ inline void ParseCigar(const vector<CigarOp> &cigar, vector<int> &blockStarts, v
 
 inline void var_processing(struct var &variant) {
 
-  unsigned int ssum;
+  float ssum;
   vector <unsigned int>::iterator sit = (variant.surrounding).begin();
   for(; sit != (variant.surrounding).end(); sit++) {
     ssum += *sit;
@@ -669,7 +669,7 @@ inline void var_processing(struct var &variant) {
     meanMis = 0.0;
     medianMis = 0.0;
   } else {
-    meanMis = ((float)ssum)/((float)surrSize);
+    meanMis = ssum/((float)surrSize);
     medianMis = CalcMedian(variant.surrounding);
   }
